Split things out a little more
