+++ parfive/__init__.py
@@ -3,6 +3,7 @@
 __version__ = '0.1.1'
 __author__ = 'Stuart Mumford <stuart@cadair.com>'
 
-from .downloader import Downloader, Results
+from .downloader import Downloader
+from .results import Results
 
 __all__ = ['Downloader', 'Results']
